Email Domain Validator/requirements.txt
aiohttp
dnspython
